double precision

+++ Job3/include/cpuRedux.h
@@ -1,6 +1,6 @@
 #ifndef __CPU_REDUX_H__
 #define __CPU_REDUX_H__
-int lowRedux(float*,unsigned int,float*,int);
+int lowRedux(double*,unsigned int,double*,int);
 #endif
 #ifndef MAT_SIZE
 #define MAT_SIZE 4
