+++ Job3/src/v0.0/MatReduxKernel.cu
@@ -1,6 +1,6 @@
 #include "kernelRedux.h"
-__global__ void  row_kernel(float *imat , float *omat){
-	__shared__ float  sdata [1024*4];
+__global__ void  row_kernel(double *imat , double *omat){
+	__shared__ double  sdata [1024*4];
 	unsigned  int tid = 4*threadIdx.x;
 	unsigned  int i = 4*(blockIdx.x*blockDim.x + threadIdx.x);
 	sdata[tid] = imat[i];
