+++ Job3/src/v0.0/cpuRedux.cpp
@@ -1,6 +1,6 @@
 #include "cpuRedux.h"
-int lowRedux(float *mat,unsigned int  length,float *ans,int flag){
-	float a1,a2,a3,a4;
+int lowRedux(double *mat,unsigned int  length,double *ans,int flag){
+	double a1,a2,a3,a4;
 	a1=a2=a3=a4=0;
 	for(unsigned int a=0;a<length;++a){
 		a1+=mat[MAT_SIZE*a];
